pyproject.toml
[tool.poetry]
name = "hoshino_bot"
version = "0.1.0"
description = ""
authors = ["菜狗 <794329728@qq.com>"]


[tool.poetry.dependencies]
python = ">=3.9,<3.11"
nonebot = "^1.8.0"
aiocqhttp = "^1.4.0"
lxml = ">=4.4.1"
pytz = ">=2019.3"
requests = ">=2.22.0"
sogou_tr_free = ">=0.0.6"
zhconv = ">=1.4.0"
Pillow = ">=6.2.1"
matplotlib = ">=3.2.0"
beautifulsoup4 = ">=4.9.0"
pygtrie = ">=2.0"
tinydb = ">=4.0"
aiohttp = ">=3.6"
numpy = "1.21.2"
jieba = "^0.42.1"
filetype = "^1.0.7"
bilibili_api = "^7.0.1"
pandas = "^1.3.2"
pinyin = "^0.4.0"
openpyxl = "^3.0.7"
playwright = "^1.14.0"
tortoise-orm = {extras = ["accel"], version = "^0.17.6"}
ujson = "^4.0.2"
types-ujson = "^0.1.1"
wordcloud = "^1.8.1"
multidict = "^5.1.0"
moviepy = "^1.0.3"
pypinyin = "^0.42.0"


[tool.poetry.dev-dependencies]



[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
